+++ src/api/Models/TermsAccepted.cs
@@ -11,7 +11,7 @@ public class TermsAccepted
     /// <summary>
     /// Gets or sets the foreign key to the <see cref="User"/> entity.
     /// </summary>
-    [Column("id_usr_fk")]
+    [Column("user_id")]
     public int UserId { get; set; }
 
     /// <summary>
@@ -22,7 +22,7 @@ public class TermsAccepted
     /// <summary>
     /// Gets or sets the foreign key to the <see cref="Term"/> entity.
     /// </summary>
-    [Column("id_tes_fk")]
+    [Column("term_id")]
     public int TermId { get; set; }
 
     /// <summary>
@@ -33,6 +33,6 @@ public class TermsAccepted
     /// <summary>
     /// Gets or sets the timestamp from the moment the terms got accepted.
     /// </summary>
-    [Column("accepted_tea")]
+    [Column("accepted")]
     public DateTime AcceptedAt { get; set; }
 }